fix: someone else's card cannot be deleted

+++ controllers/cards.js
@@ -1,3 +1,5 @@
+/* eslint-disable consistent-return */
+/* eslint-disable eqeqeq */
 const Card = require('../models/card.js');
 const handleError = require('../utils/handleError');
 
@@ -27,9 +29,19 @@ const createCard = (req, res) => {
 // GET Удаляет карточку по id
 const deleteCard = (req, res) => {
   const requestedId = req.params.id; // Запрашиваемый ID;
-  Card.findByIdAndRemove(requestedId)
-    .orFail()
-    .then(() => res.send({ message: `Карточка ${requestedId} удалена` }))
+  Card.findById(requestedId)
+    .then((requestedCard) => {
+      // Если владелец, то карточку можно удалить
+      // requestedCard.owner — object, req.user._id — String
+      if (requestedCard.owner == req.user._id) {
+        Card.findByIdAndRemove(requestedId) // Удаляем карточку
+          .orFail()
+          .then(() => res.send({ message: `Карточка ${requestedId} удалена` }))
+          .catch((err) => handleError(err, res, 'карточку. Возможно, её'));
+      } else {
+        return Promise.reject(new Error('Вы не можете удалять чужие карточки'));
+      }
+    })
     .catch((err) => handleError(err, res, 'карточку. Возможно, её'));
 };
